fix panels
